Fixed some casting.

+++ theoraplay.c
@@ -320,7 +320,7 @@ static void WorkerThread(TheoraDecoder *ctx)
                         unsigned char *yuv;
                         YuvVideoItem *item = (YuvVideoItem *) malloc(sizeof (YuvVideoItem));
                         if (item == NULL) goto cleanup;
-                        item->playms = (fps == 0) ? 0.0 : (unsigned long) ((((double) videoframes) / fps) * 1000.0);
+                        item->playms = (fps == 0) ? 0 : (unsigned int) ((((double) videoframes) / fps) * 1000.0);
                         item->fps = fps;
                         item->width = w;
                         item->height = h;
